finalized inspection imports/exports
