--- src/inspection/scope/functionExpression.ts
@@ -1,3 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
