+++ src/inspection/scope/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { tryFrom } from "./scope";
-export { InspectedIdentifier } from "./identifier";
-export { InvokeExpression, InvokeExpressionArgs, InspectedInvokeExpression } from "./invokeExpression";
+export * from "./identifier";
+export * from "./invokeExpression";
+export * from "./scope";
