+++ src/inspection/scope/invokeExpression.ts
@@ -21,7 +21,7 @@ export interface InvokeExpressionArgs {
     readonly positionArgumentIndex: number;
 }
 
-export function tryFrom(activeNode: ActiveNode, nodeIdMapCollection: NodeIdMap.Collection): InspectedInvokeExpression {
+export function inspectInvokeExpression(activeNode: ActiveNode, nodeIdMapCollection: NodeIdMap.Collection): InspectedInvokeExpression {
     const ancestors: ReadonlyArray<TXorNode> = activeNode.ancestry;
     const numAncestors: number = activeNode.ancestry.length;
     const position: Position = activeNode.position;
